userparser/requirements.txt
Django==2.2
django-compressor==2.4
django-environ==0.4.5
tabula-py==0.3.0
